Refactor batch and queue tests to enforce a strict order with mvn-surefire-plugin while allowing individual test runs through VS Code's Test Runner
