+++ pom.xml
@@ -89,9 +89,6 @@
           <version>3.5.0</version>
           <configuration>
             <argLine>--enable-preview</argLine>
-            <includes>
-              <include>**/OrderedJupiterTests.java</include>
-            </includes>
           </configuration>
         </plugin>
       </plugins>
